requirements.txt
python-dotenv>=1.0.0
firecrawl  # Add specific version once known
requests>=2.31.0